Rename GoodweLibrary to Goodwe
