+++ src/goodwe/Goodwe_inverter.py
